rearrange graphQL schema for dataTypes and resolvers to be more modular
