+++ server/schemas/comments.js
@@ -1,0 +1,7 @@
+export const typeDef = `
+  type Comments {
+    _id: ID
+    commentText: String
+    createdAt: 
+  }
+`;
