--- server/schemas/index.js
@@ -1,4 +1,0 @@
-const typeDefs = require("./typeDefs");
-const resolvers = require("./resolvers");
-
-module.exports = { typeDefs, resolvers };
